Melhoria na lógica da aplicação

+++ DevChurras.API/Controllers/ParticipanteController.cs
@@ -2,18 +2,16 @@ using DevChurras.API.Data;
 using DevChurras.API.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace DevChurras.API.Controllers
 {
     [ApiController]
-    [Route("")]
+    [Route("api/participante")]
     public class ParticipanteController : ControllerBase
     {
         [HttpGet]
-        [Route("listar_participantes")]
-        public async Task<ActionResult<List<Participante>>> Get([FromServices] DataContext context)
+        public async Task<IActionResult> GetAll([FromServices] DataContext context)
         {
             var participantes = await context.Participantes.ToListAsync();
 
@@ -21,8 +19,7 @@ namespace DevChurras.API.Controllers
         }
 
         [HttpPost]
-        [Route("adicionar_participante")]
-        public async Task<ActionResult<Participante>> Post(
+        public async Task<IActionResult> Post(
             [FromServices] DataContext context,
             [FromBody] Participante model)
         {
@@ -35,33 +32,22 @@ namespace DevChurras.API.Controllers
             return Ok(model);
         }
 
-        [HttpGet]
-        [Route("remover_participante/{id:int}")]
-        public async Task<ActionResult<Participante>> Get(
-            [FromServices] DataContext context,
-            int id)
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> Delete([FromServices] DataContext context, int id)
         {
             // Remove participante
-            var participante = await context.Participantes
-                .AsNoTracking()
-                .FirstOrDefaultAsync(p => p.ParticipanteId == id);
-
-            if (participante == null)
-                return NotFound("Participante não encontrado");
+            var participante = await context.Participantes.SingleOrDefaultAsync(p => p.Id == id);
+            if (participante == null) return NotFound("Participante não encontrado");
 
             context.Participantes.Remove(participante);
 
             // Remove convidado do participante
-            var convidado = await context.Convidados
-                .AsNoTracking()
-                .FirstOrDefaultAsync(c => c.ParticipanteId == id);
-
-            if (convidado != null)
-                context.Convidados.Remove(convidado);
+            var convidado = await context.Convidados.SingleOrDefaultAsync(c => c.ParticipanteId == id);
+            if (convidado != null) context.Convidados.Remove(convidado);
 
             await context.SaveChangesAsync();
 
-            return Ok(participante);
+            return NoContent();
         }
     }
 }
